refactor(interfaces): move UnaryFunction and OperatorFunction
moved to a separate `interfaces` file for cleaner organization
