+++ src/interfaces.ts
@@ -1,0 +1,8 @@
+import { Observable } from './Observable';
+
+export interface UnaryFunction<T, R> {
+  (source: T): R;
+}
+
+export interface OperatorFunction<T, R> extends UnaryFunction<Observable<T>, Observable<R>> {
+}
