--- src/operators/OperatorFunction.ts
@@ -1,5 +1,0 @@
-import { Observable } from '../Observable';
-
-export interface OperatorFunction<T, R> {
-  (source: Observable<T>): Observable<R>;
-}
