+++ src/operators/index.ts
@@ -1,4 +1,3 @@
-export { OperatorFunction } from './OperatorFunction';
 export { filter } from './filter';
 
 export { map } from './map';
